[Port] Ai update (#199)
* New Law Sets And Minor Fixes on AI (#1686)

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Added 4 new law boards for AI:

1- NT Agressive
2- Quarantine
3- CCTV
4- TYRANT

Now the AI ​​can see through Borg and has become a target for the Ion
Storm

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Ichai
- add: Added 4 law boards to the game: NT Agressive, Quarantine, CCTV
and TYRANT to the AI
- tweak: AI can see through borgs
- tweak: AI is now affected by ion storm

---------

Signed-off-by: Ichaie <167008606+Ichaie@users.noreply.github.com>

* moved to reserve namespace

* fix

* [Feature] Юнит, приказываю спрятать меня вон в тех техах (#440)

* Тогл болтс

* Ой

* Кролик умный

* Информационное UI окно цИИ, замена спрайта кнопок на визардские (#1433)

Добавлено информационное окно с информацией с кпк, именем цИИ и кнопками
манифеста и консоли робототехника. Теперь кнопки цИИ вынесены в
отдельные action с спрайтами визардов.
Ссылка на заказ, предложение или баг-репорт
-
[Предложение](https://discord.com/channels/901772674865455115/1358029136924377246)
-
[Предложение](https://discord.com/channels/901772674865455115/1356329554406408343)

![изображение](https://github.com/user-attachments/assets/ad44e20e-8a8a-49d8-a331-92f2386048f3)

![изображение](https://github.com/user-attachments/assets/de570f9c-9b42-4a0f-bee0-0d458a8bead4)

:cl: Cethet
- add: цИИ добавлено окно с информацией наподобие КПК
- tweak: Спрайты иконок цИИ теперь в едином стиле

* duplicates removed

* namespace changes

* more duplicates removed

* locale added

---------

Signed-off-by: Ichaie <167008606+Ichaie@users.noreply.github.com>
Co-authored-by: Ichaie <167008606+Ichaie@users.noreply.github.com>
Co-authored-by: Kai5 <68296202+Kai518@users.noreply.github.com>
Co-authored-by: Cethet <cethet159@gmail.com>